refactor : client query와 server query 분리
