+++ apis/groups/queries.server.ts
@@ -1,6 +1,6 @@
-import { getGroups } from './apis.server';
-import { Keys } from './keys';
 import { useQuery } from '@tanstack/react-query';
+import { Keys } from './keys';
+import { getGroups } from './apis.server';
 
 export const useGetGroups = (pageNum: number) => {
   return useQuery(Keys.getGroups(), () => getGroups(pageNum));
